+++ lib/ruby-next/language.rb
@@ -214,6 +214,9 @@ module RubyNext
     require "ruby-next/language/rewriters/pin_vars_pattern"
     rewriters << Rewriters::PinVarsPattern
 
+    require "ruby-next/language/rewriters/anonymous_block"
+    rewriters << Rewriters::AnonymousBlock
+
     # Put endless range in the end, 'cause Parser fails to parse it in
     # pattern matching
     require "ruby-next/language/rewriters/endless_range"